Update index.html
Had to ask Gemini to go back to a previous version of code.  The last iteration included an attempt to use React, but it wasn't set up correctly at all.   In this version, I had to get the firebaseConfig parameters from the Firebase Console and paste them into the HTML.  Let us pray!