Update index.html
Gemini needed to add something to the code to fetch the student info from the database.  Also, Gemini asked me to collect an additional parameter from the Firestore Database called firestoreCollectionName.  I believe the name for my database is "artifacts."  So I added that name along with other firebaseConfig parameters to the HTML.